Mutualise create-user and update-user
