+++ cli/create-user.php
@@ -1,21 +1,8 @@
 #!/usr/bin/php
 <?php
-require('_cli.php');
+$isUpdate = false;
+require('_update-or-create-user.php');
 
-$options = getopt('', array(
-		'user:',
-		'password:',
-		'api-password:',
-		'language:',
-		'email:',
-		'token:',
-		'no-default-feeds',
-	));
-
-if (empty($options['user'])) {
-	fail('Usage: ' . basename(__FILE__) . " --user username ( --password 'password' --api-password 'api_password'" .
-		" --language en --email user@example.net --token 'longRandomString' --no-default-feeds )");
-}
 $username = $options['user'];
 if (!FreshRSS_user_Controller::checkUsername($username)) {
 	fail('FreshRSS error: invalid username “' . $username . '”! Must be matching ' . FreshRSS_user_Controller::USERNAME_PATTERN);
@@ -30,12 +17,8 @@ echo 'FreshRSS creating user “', $username, "”…\n";
 
 $ok = FreshRSS_user_Controller::createUser($username,
 	empty($options['password']) ? '' : $options['password'],
-	empty($options['api-password']) ? '' : $options['api-password'],
-	array(
-		'language' => empty($options['language']) ? '' : $options['language'],
-		'mail_login' => empty($options['email']) ? '' : $options['email'],
-		'token' => empty($options['token']) ? '' : $options['token'],
-	),
+	empty($options['api_password']) ? '' : $options['api_password'],
+	$values,
 	!isset($options['no-default-feeds']));
 
 if (!$ok) {
